break facet test cases into two spec files
